[Chore #158722437] Avoid repitition and complexity
